+++ vaisseau/run.py
@@ -1,4 +1,0 @@
-from moteur.boucle import lancer
-
-if __name__ == "__main__":
-    lancer()